requirements.txt
langchain==0.0.208            
deeplake==3.6.7
openai==0.27.8
tiktoken==0.4.0
pydantic==1.10.10
python-dotenv==1.0.0
chainlit==0.5.0

# streamlit == 1.22.0
# streamlit_chat==0.1.1
# unstructured==0.7.12